Fixed some bugs, fonts, and visibility
There were some minor problems, such as the day string is not shown
correctly. Fonts are also available to game screen (screen that you can
choose the building). All texts and buttons now have white-transparent
background, just to make easier for testing. (we should come up with
better UI so that the player is able to know what are the buttons, etc.)